Generates a transcript query expression for matching transcripts by type, for use with getMatches or getMatchingTranscriptIds

This function generates a transcript query expression fragment which can be passed as the transcript.expression parameter of getMatches, (or the expression parameter of getMatchingTranscriptIds) in order to identify transcripts using a list of transcript types.

Usage

expressionFromTranscriptTypes(transcript.types, not = FALSE)

Arguments

transcript.types

A list of transcript types.

not

Whether to match the given IDs (FALSE), or everything except the given IDs.

Value

A transcript query expression which can be passed as the transcript.expression parameter of getMatches or the expression parameter of getMatchingTranscriptIds

See also

Examples

if (FALSE) { # \dontrun{
## Perform a search of interviews or monologues
transcript.types <- c("interview","monologue")
results <- getMatches(labbcat.url, list(segment="I"),
  transcript.expression = expressionFromTranscriptTypes(transcript.types))

## Perform a search of all transcripts that aren't word-lists.
results <- getMatches(labbcat.url, list(segment="I"),
  transcript.expression = expressionFromTranscriptTypes("wordlist", NOT=true))
} # }