Generates a layer for a given set of utterances

Generates annotations on a given layer for a given set of utterances, e.g. force-align selected utterances of a participant.

Usage

generateLayerUtterances(
  labbcat.url,
  match.ids,
  layer.id,
  collection.name = NULL,
  no.progress = FALSE
)

Arguments

labbcat.url

URL to the LaBB-CAT instance

match.ids

A vector of annotation IDs, e.g. the MatchId column, or the URL column, of a results set.

layer.id

The ID of the layer to generate.

collection.name

An optional name for the collection, e.g. the participant ID.

no.progress

TRUE to suppress visual progress bar. Otherwise, progress bar will be shown when interactive().

Value

The final status of the layer generation task.

See also

Examples

if (FALSE) { # \dontrun{
## Get all utterances of a participant
allUtterances <- getAllUtterances(labbcat.url, "AP2505_Nelson")

## Force-align the participant's utterances
generateLayerUtterances(labbcat.url, allUtterances$MatchId, "htk", "AP2505_Nelson")
} # }