Get all utterances of participants.

Identifies all utterances of a given set of participants.

getAllUtterances(
  labbcat.url,
  participant.ids,
  transcript.types = NULL,
  main.participant = TRUE,
  max.matches = NULL,
  no.progress = FALSE
)

Arguments

labbcat.url

URL to the LaBB-CAT instance

participant.ids

A list of participant IDs to identify the utterances of.

transcript.types

An optional list of transcript types to limit the results to. If null, all transcript types will be searched.

main.participant

TRUE to search only main-participant utterances, FALSE to search all utterances.

max.matches

The maximum number of matches to return, or null to return all.

no.progress

TRUE to supress visual progress bar. Otherwise, progress bar will be shown when interactive().

Value

A data frame identifying matches, containing the following columns:

SearchName A name based on the pattern -- the same for all rows

Number Row number

Transcript Name of the transcript in which the match was found

Line The start offset of the utterance/line

LineEnd The end offset of the utterance/line

MatchId A unique ID for the matching target token

Before.Match Transcript text immediately before the match

Text Transcript text of the match

Before.Match Transcript text immediately after the match

Target.word Text of the target word token

Target.word.start Start offset of the target word token

Target.word.end End offset of the target word token

Target.segment Label of the target segment (only present if the segment layer is included in the pattern)

Target.segment.start Start offset of the target segment (only present if the segment layer is included in the pattern)

Target.segment.end End offset of the target segment (only present if the segment layer is included in the pattern)

See also

Examples

if (FALSE) {
## define the LaBB-CAT URL
labbcat.url <- "https://labbcat.canterbury.ac.nz/demo/"

## get all utterances of the given participants
participant.ids <- getParticipantIds(labbcat.url)[1:3]
results <- getAllUtterances(labbcat.url, participant.ids)

## results$MatchId can be used to access results
}