Gets the annotations on the given layer of the given transcript

Returns the annotations on the given layer of the given transcript.

Usage

getAnnotations(
  labbcat.url,
  id,
  layer.id,
  max.ordinal = NULL,
  page.length = NULL,
  page.number = NULL
)

Arguments

labbcat.url

URL to the LaBB-CAT instance

id

A transcript ID (i.e. transcript name)

layer.id

A layer ID

max.ordinal

The maximum ordinal for the returned annotations. e.g. a max.ordinal of 1 will ensure that only the first annotation for each parent is returned. If max.ordinal is null, then all annotations are returned, regardless of their ordinal.

page.length

The maximum number of annotations to return, or null to return all

page.number

The zero-based page number to return, or null to return the first page

Value

A named list of annotations, with members:

id The annotation's unique ID

layerId The name of the layer it comes from

label The value of the annotation

startId The ID of the start anchor,

endId The ID of the end anchor,

parentId The ID of the parent annotation,

ordinal The ordinal of the annotation among its peers,

confidence A rating from 0-100 of the confidence of the label e.g. 10: default value, 50: automatically generated, 100: manually annotated

See also

Examples

if (FALSE) { # \dontrun{
## Get all the orthography tokens in UC427_ViktoriaPapp_A_ENG.eaf
orthography <- getAnnotations(labbcat.url, "UC427_ViktoriaPapp_A_ENG.eaf", "orthography")

## Get the first 20 orthography tokens in UC427_ViktoriaPapp_A_ENG.eaf
orthography <- getAnnotations(labbcat.url, "UC427_ViktoriaPapp_A_ENG.eaf", "orthography", 20, 0)
} # }