Lists the descriptors of all registered serializers

Returns a list of serializers, which are modules that export annotation structures as a specific file format, e.g. Praat TextGrid, plain text, etc., so the mimeType of descriptors reflects what mimeTypes can be specified for getFragments.

Usage

getSerializerDescriptors(labbcat.url)

Arguments

labbcat.url

URL to the LaBB-CAT instance

Value

A list of serializers, each including the following information:

name The name of the format.

version The installed version of the serializer module.

fileSuffixes The normal file name suffixes (extensions) of the files.

mimeType The MIME type of the format, i.e. the value to use as the mimeType parameter of getFragments

See also

Examples

if (FALSE) { # \dontrun{
## List file export formats supported
formats <- getSerializerDescriptors("https://labbcat.canterbury.ac.nz/demo/")

## can we export as plain text?
plainTextSupported <- "text/plain" %in% formats$mimeType
} # }