Upload a new transcript.

This function adds a transcript and optionally a media file to the corpus.

newTranscript(
  labbcat.url,
  transcript,
  media = NULL,
  transcript.type = NULL,
  corpus = NULL,
  episode = NULL,
  no.progress = FALSE
)

Arguments

labbcat.url

URL to the LaBB-CAT instance

transcript

The path to the transcript to upload.

media

The path to the media to upload, if any.

transcript.type

The transcript type.

corpus

The corpus to add the transcript to.

episode

The transcript's episode.

no.progress

TRUE to supress visual progress bar. Otherwise, progress bar will be shown when interactive().

Value

The ID of the new transcript in the corpus

Details

For this function to work, the credentials used to connect to the server must have at least 'edit' access.

Examples

if (FALSE) {
## define the LaBB-CAT URL
labbcat.url <- "https://labbcat.canterbury.ac.nz/demo/"

## Get attributes for new transcript
corpus <- getCorpusIds(labbcat.url)[1]
transcript.type.layer <- getLayer(labbcat.url, "transcript_type")
transcript.type <- transcript.type.layer$validLabels[[1]]

## upload transcript
newTranscript(
    labbcat.url, "my-transcript.eaf", "my-transcript.wav",
    "", transcript.type, corpus, "episode-1")
}