Accessing Data Stored in 'LaBB-CAT' Instances

'LaBB-CAT' is a web-based language corpus management system developed by the New Zealand Institute of Language, Brain and Behaviour (NZILBB) - see <https://labbcat.canterbury.ac.nz>. This package defines functions for accessing corpus data in a 'LaBB-CAT' instance. You must have at least version 20230818.1400 of 'LaBB-CAT' to use this package. For more information about 'LaBB-CAT', see Robert Fromont and Jennifer Hay (2008) <doi:10.3366/E1749503208000142> or Robert Fromont (2017) <doi:10.1016/j.csl.2017.01.004>.

Details

This package was not yet installed at build time.
Index: This package was not yet installed at build time.

'LaBB-CAT' is a web-based language corpus management system and this package provides access to data stored in a 'LaBB-CAT' instance. You must have at least version 20230818.1400 'LaBB-CAT' to use this package.

References

Robert Fromont and Jennifer Hay, "ONZE Miner: the development of a browser-based research tool", 2008 Robert Fromont, "Toward a format-neutral annotation store", 2017

See also

Author

Robert Fromont [aut, cre] (<https://orcid.org/0000-0001-5271-5487>)

Examples

if (FALSE) { # \dontrun{
## Perform a search
results <- getMatches(labbcat.url, list(segment="I")) |>
    ## Get the phonemic transcriptions for the matches
    appendLabels("phonemes")

## Get sound fragments for the matches
wav.files <- getSoundFragments(labbcat.url, results$Transcript, results$Line, results$LineEnd)
} # }