Removes an entry from a layer dictionary

This function removes an existing entry from the dictionary that manages a given layer, and updates all affected tokens in the corpus. Words can have multiple entries.

Usage

removeLayerDictionaryEntry(labbcat.url, layer.id, key, entry = NULL)

Arguments

labbcat.url

URL to the LaBB-CAT instance

layer.id

The ID of the layer with a dictionary configured to manage it.

key

The key (word) in the dictionary to remove an entry from.

entry

The value (definition) for the given key, or NULL to remove all entries for the given key.

Value

NULL if the entry was added, or a list of error messages if not.

Details

You must have edit privileges in LaBB-CAT in order to be able to use this function.

See also

Examples

if (FALSE) { # \dontrun{
## Remove a pronunciation for "robert" from the phonemes layer dictionary
removeLayerDictionaryEntry(labbcat.url, "phonemes", "robert", "'rQ-bErt")
} # }