Set the parameters of a transcript already uploaded with transcriptUpload.

The second part of a transcript upload process started by a call to transcriptUpload(), which specifies values for the parameters required to save the uploaded transcript to LaBB-CAT's database.

Usage

transcriptUploadParameters(labbcat.url, id, parameters, no.progress = FALSE)

Arguments

labbcat.url

URL to the LaBB-CAT instance

id

Upload ID returned by the prior call to transcriptUpload().

parameters

A named list where each name is the name of a parameter returned by transcriptUpload(), and the value is the parameters value.

no.progress

TRUE to suppress visual progress bar. Otherwise, progress bar will be shown when interactive().

Value

The ID of the new transcript in the corpus

Details

NB Using transcriptUpload and transcriptUploadParameters is an alternative to using newTranscript or updateTranscript.

If the response includes more parameters, then this method should be called again to supply their values.

See also

Examples

if (FALSE) { # \dontrun{
## Get attributes for new transcript
corpus <- getCorpusIds(labbcat.url)[1]
transcript.type.layer <- getLayer(labbcat.url, "transcript_type")
transcript.type <- transcript.type.layer$validLabels[[1]]

## upload transcript and its media
result <- transcriptUpload(labbcat.url, "my-transcript.eaf", "my-transcript.wav", FALSE)

## use the default parameter values
parameterValues <- list()
for(p in 1:length(parameters$name)) parameterValues[parameters$name[p]] <- parameters$value[p]

## set the upload parameters to finalise the upload
transcript.id <- transcriptUploadParameters(labbcat.url, result$id, parameterValues)
} # }