Gets annotator information.

Retrieve information about an annotator. Annotators are modules that perform different annotation tasks. This function provides information about a given annotator, for example the currently installed version of the module, what configuration parameters it requires, etc.

getAnnotatorDescriptor(labbcat.url, annotator.id)

Arguments

labbcat.url

URL to the LaBB-CAT instance.

annotator.id

ID of the annotator module.

Value

The annotator info:

annotatorId The annotators's unique ID

version The currently install version of the annotator.

info HTML-encoded description of the function of the annotator.

infoText A plain text version of $info (converted automatically).

hasConfigWebapp Determines whether the annotator includes a web-app for installation or general configuration.

configParameterInfo An HTML-encoded definition of the installation config parameters, including a list of all parameters, and the encoding of the parameter string.

configParameterInfoText A plain text version of $configParameterInfo (converted automatically).

hasTaskWebapp Determines whether the annotator includes a web-app for task parameter configuration.

taskParameterInfo An HTML-encoded definition of the task parameters, including a list of all parameters, and the encoding of the parameter string.

taskParameterInfoText A plain text version of $taskParameterInfo (converted automatically).

hasExtWebapp Determines whether the annotator includes an extras web-app which implements functionality for providing extra data or extending functionality in an annotator-specific way.

extApiInfo An HTML-encoded document containing information about what endpoints are published by the ext web-app.

extApiInfoText A plain text version of $extApiInfo (converted automatically).

See also

Examples

if (FALSE) {
## Get information about the BAS Annotator
basAnnotator <- getAnnotatorDescriptor("https://labbcat.canterbury.ac.nz/demo/", "BASAnnotator")
cat(basAnnotator$infoText)
}